Run pages router 404 test in all environments
